pyproject.toml
[tool.poetry]
name = "report_cruncher"
version = "0.1.0"
description = ""
authors = ["Report cruncher"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.1"
pandas = "^1.4.2"
prometheus-client = "^0.14.1"
structlog = "^21.5.0"
gunicorn = "^20.1.0"
openpyxl = "^3.0.9"
coverage = "^6.3.2"
pylint = "^2.13.5"
requests = "^2.27.1"
Flask-Cors = "^3.0.10"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"
retry = "^0.9.2"
confluent_kafka = "^1.8.2"
minio = "^7.1.8"
mockito = "^1.3.3"
pre-commit = "^2.20.0"
python-dotenv = "^0.20.0"
pytest-datafiles = "^2.0.1"
pytest-cov = "^3.0.0"
psycopg2-binary = "^2.9.5"
black = "^23.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mockito = "^1.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
